Saque warnings y ya esta lo de la velocidad diferente en cada direccion. Subo también otro personaje, el del principio sin el halo violeta, por las dudas.

+++ View/Personaje.h
@@ -28,7 +28,7 @@ private:
 	std::vector<Sprite*> sprites;
 	float velocidad;
 	int estado;
-	std::pair<int, int> delta;
+	std::pair<float, float> delta;
 	std::pair<float, float> ePot;
 	int serr;
 };
